Tests/LibWeb/Text/expected/Streams/init-from-fetch.txt

Was able to create a reader from a Fetch response!
